Format of data received has changed
Renaming res to result, since it may be confused with response.
Instead of res.request.uri.path
the path is now found in result.req.path
as on util.inspect ing the result object we find

[ { req:
     { domain: null,
       ...
       method: 'GET',
       path: '/teams/233393/members/saraf?access_token=fa5657bf47838731ef4ce82a929faba19803a097', //<=== this is what we want
       ...
       res: [Object] },
    res:
     { _readableState: [Object],
       readable: false,
       domain: null,
       ...

Note: I do not know if we can reliably test for the account being
an enterprise git account by looking at the path for "/api/v3"
(and the req object does not now have a "uri" property anymore.)
Needs to be tested/looked at by someone with an enterprise git account.