Testing:  getting repos - mocking GitHub with nock
Simulate a case where a user Strider Tester is registered
with github and has admin access to TWO repositories
one which belongs to him (stridertester/proj1) and one that
belongs to a team Strider Testers Union (stridertestersunion/union-proj-1)

getRepos should return an array containing the two repositories.
we are using actual responses received from github.com - as recorded
and mocked by nock to simulate.

This test was failing and we were getting three repositories.
This is because when getting the repos for the user, we also get
all the repos that he has access to from the teams of which he is a
member.

Also this was from the fix I intended to check in for
https://github.com/Strider-CD/strider/issues/846

xdissent has already checked in the fix,
I have changed the comment to explain why a de-dupe is really needed.