Removed non-working code
it is guaranteed to not work as it stands, as we do not get the uri
property in the req object at all (atleast when testing with github)
and with github enterprise - things just dont't get set up correctly
in the first place.

Also, the string /api/v3 will be a part of the uri (which is not
present in the result), however - a part of the endpoint URL and  not
a part of the path (which is present in the result) in any case.